lint: unnecessary casts

+++ native/src/context/class.rs
@@ -65,7 +65,7 @@ declare_types! {
     method transform(mut cx){
       let mut this = cx.this();
       let t = float_args(&mut cx, 0..6)?;
-      let matrix = Matrix::new_all(t[0], t[2], t[4], t[1], t[3], t[5], 0.0 as f32, 0.0 as f32, 1.0 as f32);
+      let matrix = Matrix::new_all(t[0], t[2], t[4], t[1], t[3], t[5], 0.0, 0.0, 1.0);
 
       cx.borrow_mut(&mut this, |mut this| {
         this.with_matrix(|ctm| ctm.pre_concat(&matrix) );
